string based constructor on Service
